% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_helpers.R
\name{add_crcl_cg}
\alias{add_crcl_cg}
\title{Add CRCL Column via Cockcroft-Gault}
\usage{
add_crcl_cg(
  df,
  age_col = "AGE",
  bwt_col = "BWT",
  sex_col = "SEX",
  crea_col = "CREA",
  out_col = "CRCL"
)
}
\arguments{
\item{df}{data.frame containing AGE, BWT, SEX, CREA columns.}

\item{age_col}{name of the age column (years). Default "AGE".}

\item{bwt_col}{name of the body-weight column (kg). Default "BWT".}

\item{sex_col}{name of the sex column (0=M, 1=F). Default "SEX".}

\item{crea_col}{name of the creatinine column (mg/dL). Default "CREA".}

\item{out_col}{name of the output column. Default "CRCL".}
}
\value{
data.frame with the CRCL column added.
}
\description{
Add a CRCL column to a NM-style data.frame using Cockcroft-Gault.
Column names are configurable.
}
