% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{AddDoNoTaLD}
\alias{AddDoNoTaLD}
\title{Add Dose Number, Time of Latest Dose, and Time after Latest Dose}
\usage{
AddDoNoTaLD(
  NMData,
  ID = "ID",
  TIME = "TIME",
  AMT = "AMT",
  II = "II",
  ADDL = "ADDL",
  MDV = "MDV"
)
}
\arguments{
\item{NMData}{data.frame of NONMEM dataset}

\item{ID}{character, column name for subject ID}

\item{TIME}{character, column name for time}

\item{AMT}{character, column name for dose amount}

\item{II}{character, column name for interdose interval}

\item{ADDL}{character, column name for additional doses}

\item{MDV}{character, column name for missing dependent variable flag}
}
\value{
data.frame with DoNo, ToLD, TaLD columns added
}
\description{
Adds DoNo (Dosing Occasion Number), ToLD (Time of Latest Dose), and
TaLD (Time after Latest Dose) columns to a NONMEM dataset.
}
