% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_report.R
\name{AddPage}
\alias{AddPage}
\title{Add a New Page to PDF}
\usage{
AddPage(
  Cex = 0.8,
  Header1 = "",
  Header2 = "",
  Header3 = "",
  Footer1 = "",
  Footer2 = "",
  Footer3 = "",
  PrintRowNum = FALSE,
  StartRowNum = 1
)
}
\arguments{
\item{Cex}{numeric, character expansion factor (0.8 or 0.6)}

\item{Header1}{character, left header}

\item{Header2}{character, center header}

\item{Header3}{character, right header}

\item{Footer1}{character, left footer}

\item{Footer2}{character, center footer}

\item{Footer3}{character, right footer}

\item{PrintRowNum}{logical, whether to print row numbers}

\item{StartRowNum}{integer, starting row number}
}
\description{
Creates a new page in the PDF output with optional headers and footers.
}
