% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_stats.R
\name{AllSame}
\alias{AllSame}
\title{Check if All Values are the Same}
\usage{
AllSame(Column)
}
\arguments{
\item{Column}{vector}
}
\value{
logical
}
\description{
Check if All Values are the Same
}
\keyword{internal}
