% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_parse.R
\name{BtwTagLines}
\alias{BtwTagLines}
\title{Extract Lines Between XML Tags}
\usage{
BtwTagLines(Tag1, Tag2, RawRead)
}
\arguments{
\item{Tag1}{opening tag}

\item{Tag2}{closing tag}

\item{RawRead}{character vector of lines}
}
\value{
character vector of lines between tags
}
\description{
Extract Lines Between XML Tags
}
\keyword{internal}
