% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_parse.R
\name{BtwTagVal}
\alias{BtwTagVal}
\title{Extract Value Between XML Tags}
\usage{
BtwTagVal(Tag1, Tag2, RawRead)
}
\arguments{
\item{Tag1}{opening tag}

\item{Tag2}{closing tag}

\item{RawRead}{character vector of lines}
}
\value{
character, the value between tags
}
\description{
Extract Value Between XML Tags
}
\keyword{internal}
