% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_helpers.R
\name{build_dose_records}
\alias{build_dose_records}
\title{Build NONMEM Dose Records}
\usage{
build_dose_records(EX, dose_cmt = 1L)
}
\arguments{
\item{EX}{data.frame with dose information.}

\item{dose_cmt}{default compartment for dose records lacking a CMT column.}
}
\value{
data.frame of dose records with columns
  SUBJID, DAT2, TIME, AMT, RATE, CMT, DV, MDV.
}
\description{
Build NONMEM dose records from an EX-like data.frame.
EX must contain: SUBJID, DAT2, TIME, AMT, RATE.
Optional: CMT (defaults to \code{dose_cmt}).
Output rows have MDV = 1, DV = 0.
}
