% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_helpers.R
\name{build_obs_records}
\alias{build_obs_records}
\title{Build NONMEM Observation Records}
\usage{
build_obs_records(PC)
}
\arguments{
\item{PC}{data.frame with observation information.}
}
\value{
data.frame of observation records with columns
  SUBJID, DAT2, TIME, AMT, RATE, CMT, DV, MDV.
}
\description{
Build NONMEM observation records from a PC-like data.frame.
PC must contain: SUBJID, DAT2, TIME, DV, CMT.
Output rows have AMT = 0, RATE = 0; DV that is NA or 0 -> MDV = 1.
}
