% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_helpers.R
\name{code_race}
\alias{code_race}
\title{Code RACE as Integer}
\usage{
code_race(x)
}
\arguments{
\item{x}{character or numeric vector of race values.}
}
\value{
integer vector.
}
\description{
Code RACE as integer: ASIAN -> 1, WHITE -> 2, BLACK -> 3, otherwise -> 4.
Numeric input is returned unchanged (cast to integer).
}
