% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_helpers.R
\name{code_sex}
\alias{code_sex}
\title{Code SEX as Integer}
\usage{
code_sex(x)
}
\arguments{
\item{x}{character or numeric vector of sex values.}
}
\value{
integer vector.
}
\description{
Code SEX as integer: M -> 0, F -> 1, otherwise -> 2.
Numeric input is returned unchanged (cast to integer).
}
