% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_summary.R
\name{ConnPoint}
\alias{ConnPoint}
\title{Connect Two Points with Lines}
\usage{
ConnPoint(pt1, pt2)
}
\arguments{
\item{pt1}{numeric vector of length 2, first point (x, y)}

\item{pt2}{numeric vector of length 2, second point (x, y)}
}
\description{
Draws connecting lines between parent and child nodes in flow diagram.
}
\keyword{internal}
