% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_helpers.R
\name{crcl_cg}
\alias{crcl_cg}
\title{Cockcroft-Gault Creatinine Clearance}
\usage{
crcl_cg(age, bwt, sex, crea)
}
\arguments{
\item{age}{numeric, age in years.}

\item{bwt}{numeric, body weight in kg.}

\item{sex}{integer/numeric, 0 = male, 1 = female.}

\item{crea}{numeric, serum creatinine in mg/dL.}
}
\value{
numeric vector of CRCL in mL/min.
}
\description{
Cockcroft-Gault creatinine clearance (mL/min).
}
