% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_helpers.R
\name{dat2_time_to_posix}
\alias{dat2_time_to_posix}
\title{Combine Date and Time Strings into POSIXct}
\usage{
dat2_time_to_posix(dat2, time, tz = "UTC")
}
\arguments{
\item{dat2}{character vector of dates in "YYYY-MM-DD" format.}

\item{time}{character vector of times in "HH:MM" format.}

\item{tz}{time zone. Default "UTC".}
}
\value{
POSIXct vector.
}
\description{
Combine "YYYY-MM-DD" and "HH:MM" character vectors into POSIXct.
}
