% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_ebe.R
\name{.ebe_correlation_tables}
\alias{.ebe_correlation_tables}
\title{EBE Correlation and Covariance Tables}
\usage{
.ebe_correlation_tables(tabEta, nEta, EtaNames2, OM, VarStat, FDATA)
}
\arguments{
\item{tabEta}{data.frame with ID, ETA, and covariate columns}

\item{nEta}{integer, number of ETAs}

\item{EtaNames2}{character vector, c("ID", "ETA1", "ETA2", ...)}

\item{OM}{matrix, omega matrix (nEta x nEta)}

\item{VarStat}{matrix from \code{NMVarStat()}}

\item{FDATA}{data.frame of NONMEM data}
}
\description{
Prints correlation and covariance tables comparing empirical Bayes
estimates (EBE) with the estimated omega matrix, including:
\itemize{
  \item Correlation of covariates and EBE
  \item Covariance of EBE vs omega matrix
  \item Ratios of Cov(EBE)/OM and Cor(EBE)/Cor(OM)
}
Helper function for \code{\link{nmw_report_ebe}}.
}
\keyword{internal}
