% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{ExpandDoseHist}
\alias{ExpandDoseHist}
\title{Expand Dose History with ADDL/II Records}
\usage{
ExpandDoseHist(DoseHistTab)
}
\arguments{
\item{DoseHistTab}{data.frame with columns TIME, AMT, II, ADDL}
}
\value{
data.frame with expanded dose records
}
\description{
Expands compressed dosing records (ADDL/II) into individual dose records.
}
