% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_read.R
\name{FileTag}
\alias{FileTag}
\title{Check for Tag in File}
\usage{
FileTag(FileName, Tag)
}
\arguments{
\item{FileName}{character, filename}

\item{Tag}{character, text to search for}
}
\value{
logical or NA
}
\description{
Check for Tag in File
}
\keyword{internal}
