% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_stats.R
\name{FuncDep}
\alias{FuncDep}
\title{Test Functional Dependency}
\usage{
FuncDep(NMTable, DetNameList, DepNameList)
}
\arguments{
\item{NMTable}{data.frame}

\item{DetNameList}{character vector of determinant column names}

\item{DepNameList}{character vector of dependent column names}
}
\value{
logical, TRUE if functionally dependent
}
\description{
Tests whether DepNameList columns are functionally dependent on
DetNameList columns in a NONMEM table.
}
\keyword{internal}
