% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_read.R
\name{GetCountAllTheta}
\alias{GetCountAllTheta}
\title{Get Count of All Thetas}
\usage{
GetCountAllTheta()
}
\value{
integer, total number of thetas, or NA if file not found
}
\description{
Get Count of All Thetas
}
\keyword{internal}
