% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_read.R
\name{GetCountOmega}
\alias{GetCountOmega}
\title{Get Count of Unfixed Omega Elements}
\usage{
GetCountOmega()
}
\value{
integer, or NA if file not found
}
\description{
Get Count of Unfixed Omega Elements
}
\keyword{internal}
