% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_read.R
\name{GetCountPara}
\alias{GetCountPara}
\title{Get Count of Parameters}
\usage{
GetCountPara()
}
\value{
integer, total number of estimated parameters, or NA if file not found
}
\description{
Get Count of Parameters
}
