% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_read.R
\name{GetCountUnfixedEps}
\alias{GetCountUnfixedEps}
\title{Get Count of Unfixed Epsilons}
\usage{
GetCountUnfixedEps()
}
\value{
integer, or NA if file not found
}
\description{
Get Count of Unfixed Epsilons
}
\keyword{internal}
