% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_read.R
\name{GetEstMethod}
\alias{GetEstMethod}
\title{Get Estimation Method}
\usage{
GetEstMethod()
}
\value{
character, abbreviation of the estimation method
}
\description{
Identifies the estimation method from the .ext file header.
}
