% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_summary.R
\name{GetModelNames}
\alias{GetModelNames}
\title{Get Model Names from Folder Names}
\usage{
GetModelNames(vName)
}
\arguments{
\item{vName}{character vector of folder names}
}
\value{
character vector of model names
}
\description{
Extracts model names by splitting folder names on period.
}
\keyword{internal}
