% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report.R
\name{GetNRecFromXML}
\alias{GetNRecFromXML}
\title{Number of Data Records from NONMEM XML/Report Lines}
\usage{
GetNRecFromXML(XML)
}
\arguments{
\item{XML}{character vector of NONMEM output lines}
}
\value{
integer record count
}
\description{
Extracts the "NO. OF DATA RECS IN DATA SET" count. Shared between
report_resid.R and report_input.R.
}
\keyword{internal}
