% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_read.R
\name{GetOFV}
\alias{GetOFV}
\title{Get Objective Function Value}
\usage{
GetOFV()
}
\value{
numeric, the objective function value
}
\description{
Retrieves the final OFV from the .ext file in the current directory.
}
