% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_read.R
\name{GetProbVal}
\alias{GetProbVal}
\title{Get PROBLEM Statement Value}
\usage{
GetProbVal(Tag)
}
\arguments{
\item{Tag}{character, the tag to search for in FCON}
}
\value{
character, the extracted value
}
\description{
Get PROBLEM Statement Value
}
\keyword{internal}
