% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_summary.R
\name{GetReason}
\alias{GetReason}
\title{Get Termination Reason from PRINT.OUT}
\usage{
GetReason(folderName)
}
\arguments{
\item{folderName}{character, optional folder to set as working directory}
}
\value{
character, termination reason code
}
\description{
Get Termination Reason from PRINT.OUT
}
\keyword{internal}
