% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_stats.R
\name{IntStat}
\alias{IntStat}
\title{Integer Variable Statistics}
\usage{
IntStat(NMTable, VarStat)
}
\arguments{
\item{NMTable}{data.frame}

\item{VarStat}{matrix from NMVarStat}
}
\description{
Prints unique values for all integer variables.
}
\keyword{internal}
