% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_math.R
\name{LogDV}
\alias{LogDV}
\title{Check if Model Uses Log-Transformed DV (from Control File)}
\usage{
LogDV(CtlFileName)
}
\arguments{
\item{CtlFileName}{character, path to NONMEM control file}
}
\value{
logical, TRUE if LOG(F) is found after $PRED or $ERROR
}
\description{
Check if Model Uses Log-Transformed DV (from Control File)
}
