% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_math.R
\name{LogDV2}
\alias{LogDV2}
\title{Check if Model Uses Log-Transformed DV (from FSUBS)}
\usage{
LogDV2()
}
\value{
logical, TRUE if LOG(F) is found after SUBROUTINE PRED or ERROR in FSUBS
}
\description{
Check if Model Uses Log-Transformed DV (from FSUBS)
}
