% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_string.R
\name{MatchEnd}
\alias{MatchEnd}
\title{Match End of String}
\usage{
MatchEnd(vName0, End0, IgnoreCase = TRUE)
}
\arguments{
\item{vName0}{character vector of names}

\item{End0}{character string to match at end}

\item{IgnoreCase}{logical, whether to ignore case}
}
\value{
logical vector
}
\description{
Match End of String
}
\keyword{internal}
