% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report.R
\name{MergeIDStatOFV}
\alias{MergeIDStatOFV}
\title{Merge IDStat with Individual OFV from PHI}
\usage{
MergeIDStatOFV(FDATA, PHI)
}
\arguments{
\item{FDATA}{data.frame of NONMEM data (from \code{ReadFDATA()})}

\item{PHI}{data.frame from the .phi file, must contain ID and OBJ columns}
}
\value{
list with components:
\describe{
  \item{IDStat3}{data.frame, merged individual statistics with OFV}
  \item{VarStat}{matrix, variable statistics from \code{NMVarStat()}}
  \item{IDStat}{list, individual statistics from \code{NMIDStat()}}
}
}
\description{
Computes variable and individual statistics from NONMEM data,
then merges individual OFV values from the PHI file. This is a
common pattern used in \code{report_ofv.R} and \code{report_resid.R}.
}
\keyword{internal}
