\name{MinUtil}
\alias{SqrtInvCov}
\alias{Mx}
\alias{mat2ltv}
\alias{ltv2mat}
\alias{mat2utv}
\alias{utv2mat}
\alias{ScaleVar}
\alias{DesclVar}
\alias{DECN}
\title{Internal Min Util Functions}
\description{Internal Min Util functions}
\details{These are not to be called by the user.}
\keyword{internal}
