% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_math.R
\name{OBJfpost}
\alias{OBJfpost}
\title{Post-Hoc OFV and CWRES Calculation}
\usage{
OBJfpost(ctlname, sdtab)
}
\arguments{
\item{ctlname}{character, control file name (without extension)}

\item{sdtab}{data.frame, standard output table}
}
\value{
list with OFV, OFVi, CWRES
}
\description{
Calculates individual OFV and CWRES using METHOD=COND or METHOD=ZERO.
}
\keyword{internal}
