% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_summary.R
\name{Outline}
\alias{Outline}
\title{Draw Model Development Flow Diagram}
\usage{
Outline(MDL, MDLName, out.start = "0", out.end = "zzzzzzzz", Target = "PDF")
}
\arguments{
\item{MDL}{data.frame from SumOut}

\item{MDLName}{character, model development name for title}

\item{out.start}{character, filter start}

\item{out.end}{character, filter end}

\item{Target}{character, "PDF" or other}
}
\description{
Creates a graphical model development flow diagram from SumOut results.
}
