% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pairs.R
\name{PanelChar}
\alias{PanelChar}
\title{Character Panel for Pairs Plot}
\usage{
PanelChar(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\description{
Character Panel for Pairs Plot
}
\keyword{internal}
