% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pairs.R
\name{PanelCor}
\alias{PanelCor}
\title{Correlation Panel for Pairs Plot}
\usage{
PanelCor(x, y, digits = 2, prefix = "", cex.cor)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{digits}{integer, number of digits for correlation}

\item{prefix}{character, prefix for text}

\item{cex.cor}{numeric, character expansion for correlation}
}
\description{
Correlation Panel for Pairs Plot
}
\keyword{internal}
