% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pairs.R
\name{PanelHist}
\alias{PanelHist}
\title{Histogram Panel for Pairs Plot}
\usage{
PanelHist(x, ...)
}
\arguments{
\item{x}{numeric vector}

\item{...}{additional arguments}
}
\description{
Histogram Panel for Pairs Plot
}
\keyword{internal}
