% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_parse.R
\name{ParseItemName}
\alias{ParseItemName}
\title{Parse Data Item Names from NONMEM Output}
\usage{
ParseItemName(RawRead)
}
\arguments{
\item{RawRead}{character vector of lines from NONMEM output}
}
\value{
character vector of item names
}
\description{
Parse Data Item Names from NONMEM Output
}
\keyword{internal}
