% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_parse.R
\name{ParseItemWidth}
\alias{ParseItemWidth}
\title{Parse Fixed-Width Format Item Widths from NONMEM Output}
\usage{
ParseItemWidth(RawRead)
}
\arguments{
\item{RawRead}{character vector of lines from NONMEM output}
}
\value{
list with LineCount and Widths
}
\description{
Parse Fixed-Width Format Item Widths from NONMEM Output
}
\keyword{internal}
