% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_parse.R
\name{ParseOut}
\alias{ParseOut}
\title{Parse Value After Tag from NONMEM Output}
\usage{
ParseOut(TagString, RawRead)
}
\arguments{
\item{TagString}{character, the tag to search for}

\item{RawRead}{character vector of lines from NONMEM output}
}
\value{
character, the value after the tag
}
\description{
Extracts the value following a tag string (after the colon) in NONMEM output lines.
}
