% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_report.R
\name{PrepPDF}
\alias{PrepPDF}
\title{Prepare PDF Output}
\usage{
PrepPDF(FileName, Paper = "letter", FontFamily = "Courier")
}
\arguments{
\item{FileName}{character, output PDF filename}

\item{Paper}{character, paper size (default "letter")}

\item{FontFamily}{character, font family (default "Courier")}
}
\description{
Opens a PDF device for NONMEM report generation.
}
