% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_report.R
\name{PrinMTxt}
\alias{PrinMTxt}
\title{Print Multiple Lines of Text}
\usage{
PrinMTxt(
  MTxt,
  Cex = 0.8,
  Header1 = "",
  Header2 = "",
  Header3 = "",
  Footer1 = "",
  Footer2 = "",
  Footer3 = "",
  PrintRowNum = FALSE
)
}
\arguments{
\item{MTxt}{character vector of text lines}

\item{Cex}{numeric, character expansion factor}

\item{Header1}{character, left header}

\item{Header2}{character, center header}

\item{Header3}{character, right header}

\item{Footer1}{character, left footer}

\item{Footer2}{character, center footer}

\item{Footer3}{character, right footer}

\item{PrintRowNum}{logical, whether to print row numbers}
}
\description{
Prints a character vector of text lines across multiple pages as needed.
}
