% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report.R
\name{ReadEXTFile}
\alias{ReadEXTFile}
\title{Read EXT File from Current Directory}
\usage{
ReadEXTFile()
}
\value{
data.frame or NULL if file not found
}
\description{
Read EXT File from Current Directory
}
\keyword{internal}
