% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{RemoveNA}
\alias{RemoveNA}
\title{Remove Rows with NA Values}
\usage{
RemoveNA(RawData)
}
\arguments{
\item{RawData}{data.frame}
}
\value{
data.frame with rows containing NA removed
}
\description{
Remove Rows with NA Values
}
