% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{RenCol}
\alias{RenCol}
\title{Rename Columns}
\usage{
RenCol(ColList, OldCol, NewCol)
}
\arguments{
\item{ColList}{character vector of column names}

\item{OldCol}{old column name}

\item{NewCol}{new column name}
}
\value{
character vector with renamed columns
}
\description{
Rename Columns
}
\keyword{internal}
