% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{RmvCol}
\alias{RmvCol}
\title{Remove Columns from Table}
\usage{
RmvCol(Tab, OldCols)
}
\arguments{
\item{Tab}{data.frame}

\item{OldCols}{character vector of column names to remove}
}
\value{
data.frame with specified columns removed
}
\description{
Remove Columns from Table
}
\keyword{internal}
