% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{RmvZero}
\alias{RmvZero}
\title{Remove Zero Rows/Cols from Symmetric Matrix}
\usage{
RmvZero(SymmMat)
}
\arguments{
\item{SymmMat}{symmetric matrix}
}
\value{
matrix with zero rows/columns removed
}
\description{
Remove Zero Rows/Cols from Symmetric Matrix
}
\keyword{internal}
