% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_tests.R
\name{run.p}
\alias{run.p}
\title{Run Test Probability}
\usage{
run.p(m, n, r)
}
\arguments{
\item{m}{count of fewer species (minimum value = 0)}

\item{n}{count of more frequent species (minimum value = 1)}

\item{r}{count of runs (minimum value = 1)}
}
\value{
numeric, probability
}
\description{
Calculates P(Run count <= r | m, n) for the runs test.
}
\keyword{internal}
