% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_tests.R
\name{run.test.nm}
\alias{run.test.nm}
\title{Runs Test for NONMEM Residuals}
\usage{
run.test.nm(RES)
}
\arguments{
\item{RES}{numeric vector of residuals}
}
\value{
numeric, two-sided p-value
}
\description{
Performs a runs test on residuals. Zeros are omitted.
}
