% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_summary.R
\name{SumOut1}
\alias{SumOut1}
\title{Summarize Single NONMEM Run}
\usage{
SumOut1(ModelName)
}
\arguments{
\item{ModelName}{character, model name}
}
\value{
data.frame with one row of summary metrics
}
\description{
Collects key metrics from a single NONMEM run folder.
}
\keyword{internal}
