% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report.R
\name{CalcTaLDForReport}
\alias{CalcTaLDForReport}
\title{Calculate TaLD (Time after Latest Dose) for Report Generation}
\usage{
CalcTaLDForReport(FDATA, nRec)
}
\arguments{
\item{FDATA}{data.frame of NONMEM data}

\item{nRec}{integer, number of records}
}
\value{
list with ToLD, TaLD, DOCC, NewID vectors
}
\description{
Shared logic between report_pred.R and report_resid.R.
}
\keyword{internal}
