% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report.R
\name{CountEXTParams}
\alias{CountEXTParams}
\title{Count THETA, ETA, EPS Parameters from EXT Data Frame}
\usage{
CountEXTParams(EXT)
}
\arguments{
\item{EXT}{data.frame from .ext file (already read)}
}
\value{
list with nTheta, nEta, nEps
}
\description{
Count THETA, ETA, EPS Parameters from EXT Data Frame
}
\keyword{internal}
