% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report.R
\name{MergeIDStatOFV}
\alias{MergeIDStatOFV}
\title{Merge IDStat with Individual OFV from PHI}
\usage{
MergeIDStatOFV(FDATA, PHI)
}
\arguments{
\item{FDATA}{data.frame of NONMEM data}

\item{PHI}{data.frame from .phi file}
}
\value{
list with IDStat3 (merged), VarStat, IDStat
}
\description{
Common pattern used in report_ofv.R and report_resid.R.
}
\keyword{internal}
