% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_output.R
\name{nmw_report_output}
\alias{nmw_report_output}
\title{Generate NONMEM Output Report (SB-Output.PDF)}
\usage{
nmw_report_output(run_dir = getwd())
}
\arguments{
\item{run_dir}{character, path to the NONMEM run directory (default: current directory)}
}
\description{
Includes the NONMEM PRINT.OUT file content in a PDF report.
}
