% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report.R
\name{OFV_SCREEN_LAYOUT}
\alias{OFV_SCREEN_LAYOUT}
\title{OFV Report Split Screen Layout}
\usage{
OFV_SCREEN_LAYOUT
}
\description{
Layout matrix for the S1-OFV report's split screen display.
}
\keyword{internal}
