% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report.R
\name{PlotDistribution}
\alias{PlotDistribution}
\title{Plot Distribution Panel (Summary + Histogram + QQ)}
\usage{
PlotDistribution(
  var.data,
  var.name,
  show_shrinkage = FALSE,
  omega_var = NULL,
  show_ttest = FALSE,
  show_cv = FALSE
)
}
\arguments{
\item{var.data}{numeric vector of data}

\item{var.name}{character, variable name for labeling}

\item{show_shrinkage}{logical, whether to show shrinkage}

\item{omega_var}{numeric, omega variance for shrinkage calc}

\item{show_ttest}{logical, whether to show t-test p-value}

\item{show_cv}{logical, whether to show CV percent}
}
\description{
Common 3-panel distribution plot used in report_ebe.R,
report_indipk.R, and report_resid.R.
}
\keyword{internal}
