% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_summary.R
\name{SumOut}
\alias{SumOut}
\title{Summarize Multiple NONMEM Runs}
\usage{
SumOut(FileExt = ".CTL", RunExt = ".R75", OutExt = ".OUT")
}
\arguments{
\item{FileExt}{character, control file extension}

\item{RunExt}{character, run folder extension}

\item{OutExt}{character, output file extension}
}
\value{
data.frame with summary of all runs
}
\description{
Scans subdirectories matching RunExt and collects summary metrics
for model development flow analysis.
}
